+++ BEAM/Image/Skia/SkiaImage.cs
@@ -24,7 +24,7 @@ public abstract class SkiaImage<T> : IContiguousImage<T>, IDisposable, ISkiaImag
 
   public ImageMemoryLayout Layout { get; init; }
 
-  public T this[int x, int y, int z] => GetValue(Layout.Flatten(x, y, GetColor(z)));
+  public T this[long x, long y, int z] => GetValue(Layout.Flatten(x, y, GetColor(z)));
   public T this[long i] => GetValue(i);
 
   public double GetAsDouble(long i)
@@ -32,7 +32,7 @@ public abstract class SkiaImage<T> : IContiguousImage<T>, IDisposable, ISkiaImag
     return (double) Convert.ChangeType(this[i], typeof(double))!;
   }
 
-  public double GetAsDouble(int x, int y, int z)
+  public double GetAsDouble(long x, long y, int z)
   {
     return (double) Convert.ChangeType(this[x, y, z], typeof(double))!;
   }
